+++ src/components/material/material-table/index.ts
@@ -1,0 +1,1 @@
+export * from './material-table';
