+++ src/components/order/order-table/index.ts
@@ -1,0 +1,1 @@
+export * from './order-table';
